transactions-service/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
attrs==24.1.0
dnspython==2.6.1
frozenlist==1.4.1
html5tagger==1.3.0
httptools==0.6.1
idna==3.7
motor==3.5.1
multidict==6.0.5
pymongo==4.8.0
python-dotenv==1.0.1
sanic==24.6.0
sanic-routing==23.12.0
tracerite==1.1.1
typing_extensions==4.12.2
websockets==12.0
yarl==1.9.4
